fix: change footer Contact us links, change menu name
